Fix DrawableTexture invalid texture address
On Finalize the texture gets a new address which has to be updated
through out all the references. Now the DrawableTexture returns a
pointer to the pointer of the texture which has always the correct
value.

SpriteComponent is able to take a pointer to a pointer to a texture and
on each GetTexture call resolve the address and return the usual
SDL_Texture*