Exposing a C interface from the shared library
In this patch we start taking back some of the useful stuff from the old
engine.

We also build a shared library, with a C interface defined by the
headers in the "include" folder. In this way we can easily write a
wrapper in any other language in the future (C# maybe?)

It would be nice to expose a C++ interface too, maybe the same used
internally (?)